pyproject.toml
[project]
name = "network-com-audit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "paramiko>=3.5.1",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "textfsm>=2.1.0",
]
